requirements.txt
flask==2.3.2
pyfiglet==0.8.post1
rich==13.4.2
dnspython==2.4.2